- BUG: Filemanager when showing a page when no files are present
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@3540 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

Former-commit-id: 5ee73c1158d3f63744e3f57caf00c1803cdf3862
Former-commit-id: 4d672f559213e178034b1cbbb7a6bd54fc38287b
Former-commit-id: dfd36ccb28cf4f9aca5a0ac44ab7ebd7171ddf5b [formerly 222a24232f46e3f1db79b52f57085d779dcafbe6]
Former-commit-id: fc5cc2620de3e04629a7a4b2cdabb08665408823

+++ sys/flexyadmin/controllers/admin/Filemanager.php
@@ -54,7 +54,7 @@ class Filemanager extends AdminController {
 			$idFile=el('current',$args);
 			$info=el('info',$args);
 			// $sub=el('sub',$args);
-			$offset=el('offset',$args,0);
+			$offset=(int) el('offset',$args,0);
 			$order=el('order',$args, $this->cfg->get('cfg_media_info',$path,'str_order','name') );
 			$search=el('search',$args,'');
 		}
